+++ packages/implementations/react/src/atomic/Select/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 Use a select box when a user needs to select one or more than one option from a list.
 
-## 2024-06-11 `1.0.0`
+## 2024-06-12 `1.0.0`
 
 #### 🎉 New features
 
-- Added `children`, `htmlFor` and `hidden` properties to the component. ([#55](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/55) by [@junior](https://git.rarolabs.com.br/junior))
-- Added stories on Component. ([#55](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/55) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `onChange`, `value`, `options`, `children`, `size` and `appearance` properties to the component. ([#57](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/57) by [@junior](https://git.rarolabs.com.br/junior))
+- Added stories on Component. ([#57](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/57) by [@junior](https://git.rarolabs.com.br/junior))
 
 <!-- #### 🛠 Breaking changes -->